# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CLISENSA API
  description: API REST de la plataforma multi-hospital de citas médicas CLISENSA.
  version: 0.5.0
servers:
  - url: http://localhost:4000/api
    description: Desarrollo local
...
paths:
...
  /hospital/servicios:
    get:
      tags: [Servicios]
      summary: Listar servicios de un hospital
      parameters:
        - in: query
          name: hospitalId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Catálogo de servicios }
    post:
      tags: [Servicios]
      summary: Crear servicio en el hospital del personal autenticado
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, duracionMin, precio]
              properties:
                nombre: { type: string }
                duracionMin: { type: integer, minimum: 5 }
                precio: { type: number, minimum: 0 }
      responses:
        "201": { description: Servicio creado }
        "403": { description: Rol no autorizado }
...
  /medicos/{id}/horarios:
    post:
      tags: [Medicos]
      summary: Agregar horario base a un médico (rol administrativo)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hora]
              properties:
                hora: { type: string, pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$", example: "09:00" }
      responses:
        "201":
          description: Horario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HorarioBase"
        "400": { description: La hora falta o tiene formato inválido }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol administrativo }
        "404": { description: Médico no encontrado }
        "409": { description: El horario ya existe }
  /medicos/{id}/horarios/{horarioId}:
    delete:
      tags: [Medicos]
      summary: Quitar horario base de un médico (rol administrativo)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: horarioId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Horario eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol administrativo }
        "404": { description: Horario no encontrado para ese médico }
  /horarios/disponibles:
    get:
      tags: [Horarios]
      summary: Consultar horarios libres de un médico en una fecha
      parameters:
        - in: query
          name: medicoId
          required: true
          schema: { type: integer }
        - in: query
          name: fecha
          required: true
          schema: { type: string, format: date, example: "2026-09-30" }
        - in: query
          name: servicioId
          schema: { type: integer }
      responses:
        "200":
          description: Horarios disponibles
          content:
            application/json:
              schema:
                type: object
                properties:
                  medicoId: { type: integer }
                  fecha: { type: string, format: date }
                  disponibles:
                    type: array
                    items: { type: string, example: "09:00" }
        "400": { description: Parámetros requeridos faltantes o inválidos }
  /horarios:
    get:
      tags: [Horarios]
      summary: Consultar disponibilidad semanal del médico autenticado
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Bloques semanales configurados }
    post:
      tags: [Horarios]
      summary: Reemplazar disponibilidad semanal del médico autenticado
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [disponibilidad]
              properties:
                disponibilidad:
                  type: array
                  items:
                    type: object
                    required: [diaSemana, horaInicio, horaFin]
                    properties:
                      diaSemana: { type: integer, minimum: 0, maximum: 6 }
                      horaInicio: { type: string, example: "08:00" }
                      horaFin: { type: string, example: "12:00" }
      responses:
        "200": { description: Disponibilidad guardada }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HorarioBase:
      type: object
      required: [id, medicoId, hora]
      properties:
        id: { type: integer }
        medicoId: { type: integer }
        hora: { type: string, example: "09:00" }
...